requirements/base.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-storages==1.14.2
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7